requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
django-debug-toolbar==1.0.1
